flask/microblog-login/requirements.txt
Flask==0.12.2
Flask-wtf==0.14.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
sqlalchemy-migrate